# examples/orders-api/spec/src/main/resources/openapi/orders-api.yaml
openapi: 3.0.3
info:
  title: Orders API
  version: 1.0.0
  description: |
    Sample contract used to exercise contract-first tooling: generated Java models,
    Spring Boot 4 declarative HTTP clients, server API interfaces and spec-conformance tests.

    Every operation requires an API key passed in the `X-API-Key` header.
    Errors are returned as RFC 9457 problem details (`application/problem+json`).
  contact:
    name: Orders Platform Team
    email: orders-platform@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: http://localhost:8080/api/v1
    description: Local development

security:
  - ApiKeyAuth: []

tags:
  - name: orders
    description: Create, inspect and manage customer orders.
  - name: catalog
    description: Read-only product catalog lookups.

paths:
  /orders:
    get:
      tags: [orders]
      operationId: listOrders
      summary: List orders
      description: Returns a page of orders, optionally filtered by status, customer, creation time and tags.
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: status
          in: query
          description: Only return orders in this status.
          required: false
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: customerId
          in: query
          description: Only return orders placed by this customer.
          required: false
          schema:
            type: string
            format: uuid
        - name: createdAfter
          in: query
          description: Only return orders created strictly after this instant.
          required: false
          schema:
            type: string
            format: date-time
        - name: tag
          in: query
          description: Only return orders carrying every one of these tags. Repeat the parameter for multiple tags.
          required: false
          style: form
          explode: true
          schema:
            type: array
            maxItems: 10
            items:
              type: string
              minLength: 1
              maxLength: 32
        - name: page
          in: query
          description: Zero-based page index.
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Page size.
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: A page of orders.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [orders]
      operationId: createOrder
      summary: Create an order
      description: |
        Creates a new order in `PENDING` status. The request must carry an `Idempotency-Key`.
        Re-submitting the same key with an identical payload replays the original `201` response
        (same order, same `Location`); re-using the key with a different payload is rejected with `409 Conflict`.
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: Idempotency-Key
          in: header
          description: Client-generated key that makes the creation idempotent.
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: The order was created.
          headers:
            Location:
              description: URL of the newly created order.
              schema:
                type: string
                format: uri
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  /orders/{orderId}:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    get:
      tags: [orders]
      operationId: getOrder
      summary: Get an order by id
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: The order.
          headers:
            ETag:
              description: Entity tag derived from the order version.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [orders]
      operationId: replaceOrder
      summary: Replace an order
      description: Replaces the mutable part of a `PENDING` order. Orders in any other status are immutable.
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: If-Match
          in: header
          description: Optimistic-locking guard; must equal the current `version` of the order.
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderRequest'
      responses:
        '200':
          description: The updated order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    patch:
      tags: [orders]
      operationId: patchOrder
      summary: Partially update an order
      description: Updates only the fields present in the request body. Absent fields are left untouched.
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderPatch'
      responses:
        '200':
          description: The updated order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [orders]
      operationId: cancelOrder
      summary: Cancel an order
      description: Cancels an order that has not been shipped yet. Cancelling an already cancelled order is a no-op.
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: reason
          in: query
          description: Free-text reason recorded with the cancellation.
          required: false
          schema:
            type: string
            maxLength: 200
      responses:
        '204':
          description: The order was cancelled.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /orders/{orderId}/submit:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    post:
      tags: [orders]
      operationId: submitOrder
      summary: Submit an order for fulfilment
      description: Transitions a `PENDING` order to `SUBMITTED`. Has no request body.
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: The submitted order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /catalog/products/{sku}:
    get:
      tags: [catalog]
      operationId: getProduct
      summary: Look up a product by SKU
      parameters:
        - name: sku
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Sku'
        - name: Accept-Language
          in: header
          description: Preferred language for the product name, as a BCP 47 tag.
          required: false
          schema:
            type: string
            default: en
      responses:
        '200':
          description: The product.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Static API key issued to each client application.

  parameters:
    OrderId:
      name: orderId
      in: path
      required: true
      description: Order identifier.
      schema:
        type: string
        format: uuid
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Correlation id echoed back in the response. Generated by the server when absent.
      schema:
        type: string
        format: uuid

  headers:
    XRequestId:
      description: Correlation id of the request.
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: The request is syntactically invalid or violates a constraint.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: The API key is missing or not recognised.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: No resource exists with the given identifier.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: The request conflicts with the current state of the resource.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    PreconditionFailed:
      description: The `If-Match` precondition does not hold.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnprocessableEntity:
      description: The request is well-formed but cannot be fulfilled (for example an unknown SKU or mixed currencies).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

  schemas:
    Sku:
      type: string
      description: Stock keeping unit, upper-case alphanumerics and dashes.
      pattern: '^[A-Z0-9]+(-[A-Z0-9]+)*$'
      minLength: 3
      maxLength: 32
      example: WIDGET-BLUE-L

    OrderStatus:
      type: string
      description: Lifecycle state of an order.
      enum: [PENDING, SUBMITTED, SHIPPED, CANCELLED]

    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          description: Decimal amount in major units.
          minimum: 0
          example: 19.99
        currency:
          type: string
          description: ISO 4217 currency code.
          pattern: '^[A-Z]{3}$'
          example: EUR

    Address:
      type: object
      required: [line1, city, postalCode, countryCode]
      properties:
        line1:
          type: string
          minLength: 1
          maxLength: 120
        line2:
          type: string
          maxLength: 120
        city:
          type: string
          minLength: 1
          maxLength: 80
        postalCode:
          type: string
          minLength: 1
          maxLength: 16
        countryCode:
          type: string
          description: ISO 3166-1 alpha-2 country code.
          pattern: '^[A-Z]{2}$'

    OrderLine:
      type: object
      required: [sku, quantity, unitPrice]
      properties:
        sku:
          $ref: '#/components/schemas/Sku'
        quantity:
          type: integer
          format: int32
          minimum: 1
          maximum: 1000
        unitPrice:
          $ref: '#/components/schemas/Money'

    CreateOrderRequest:
      type: object
      required: [customerId, lines, shippingAddress]
      properties:
        customerId:
          type: string
          format: uuid
        lines:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderLine'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        notes:
          type: string
          maxLength: 500
        tags:
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 32

    UpdateOrderRequest:
      type: object
      description: Full replacement of the mutable part of an order.
      required: [lines, shippingAddress]
      properties:
        lines:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderLine'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        notes:
          type: string
          maxLength: 500
        tags:
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 32

    OrderPatch:
      type: object
      description: |
        Partial update with JSON Merge Patch semantics (RFC 7396): absent properties are left untouched,
        `notes: null` clears the notes, and an empty `tags` array removes all tags.
      properties:
        shippingAddress:
          $ref: '#/components/schemas/Address'
        notes:
          type: string
          nullable: true
          maxLength: 500
        tags:
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 32

    Order:
      type: object
      required: [id, customerId, status, lines, shippingAddress, total, createdAt, updatedAt, version]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        customerId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        lines:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderLine'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        notes:
          type: string
          maxLength: 500
        tags:
          type: array
          uniqueItems: true
          items:
            type: string
        cancellationReason:
          type: string
          maxLength: 200
          readOnly: true
          description: Reason supplied when the order was cancelled; absent otherwise.
        total:
          $ref: '#/components/schemas/Money'
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        version:
          type: integer
          format: int64
          description: Monotonically increasing revision used for optimistic locking.
          readOnly: true

    OrderPage:
      type: object
      required: [items, page, size, totalElements]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0

    Product:
      type: object
      required: [sku, name, price, inStock]
      properties:
        sku:
          $ref: '#/components/schemas/Sku'
        name:
          type: string
        price:
          $ref: '#/components/schemas/Money'
        inStock:
          type: boolean

    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the problem type.
          default: about:blank
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI identifying this specific occurrence.
        errors:
          type: array
          description: Field-level validation errors, present for `400` responses.
          items:
            $ref: '#/components/schemas/FieldError'

    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          description: JSON pointer-ish path of the offending field or parameter name.
        message:
          type: string
        rejectedValue:
          type: string
          description: String rendering of the rejected value, when safe to echo.